requirements.txt
pytorch-lightning==0.7.5
Pillow==6.2.1
torch==1.3.1
matplotlib
jupyter
pandas
sklearn